requirements.txt
attrs==21.2.0
autopep8==1.6.0
cattrs==1.8.0
cdk8s==1.1.17
cdk8s-plus-22==1.0.0b24
constructs==3.3.161
jsii==1.41.0
publication==0.0.3
pycodestyle==2.8.0
python-dateutil==2.8.2
PyYAML==6.0
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.2
